script_function is now passed as a shared pointer Fixed check node node again... Added ability to enable and disable controls (will add script functions soon) Fixed screen not resizing to custom target size Fixed vertex batch not rotating rectagular quads correctly

+++ include/engine/time.hpp
@@ -140,7 +140,7 @@ class frame_clock
 {
 	clock mFps_clock;
 	clock mDelta_clock;
-	unsigned int mFrames;
+	size_t mFrames;
 	float mFps;
 	float mInterval;
 	float mDelta;
